refactor(tui): extract modal panels and share resolution
The share invite, rename prompt, and delete-tab confirmation move into
tui::render::modals; resolving a local ticket and reporting a copy move
into tui::share.
