+++ src/tui/render/mod.rs
@@ -2,4 +2,5 @@
 
 pub(in crate::tui) mod agents;
 pub(in crate::tui) mod footer;
+pub(in crate::tui) mod modals;
 pub(in crate::tui) mod vt;
